requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
openai
python-multipart>=0.0.6
starlette>=0.27.0
tenacity>=8.2.3
requests>=2.31.0
asyncio>=3.4.3
black>=23.10.0
isort>=5.12.0
mypy>=1.6.1
flake8>=6.1.0
pyyaml>=6.0
httpx>=0.25.0

# Database clients
pymongo>=4.5.0
motor>=3.3.1
qdrant-client>=1.6.0
neo4j>=5.13.0

# AWS
boto3>=1.28.0
anthropic[bedrock]

# Reasoning agent
agno

motor
python-hcl2
fastapi_mcp
mcp_proxy
mcp==1.7.0